chore: update statsGl y stats
